Update README.md
Manual changes to the .md file - since I have to export this in a weird way from RStudio because pandoc doesn't play well with work's network drives, all the image src were hard-coded to my desktop so they wouldn't display. Trimmed the path to start with README_files/ and that fixes the whole thing.