Update README.md
changed the path to the logo to the vignettes folder for the readme, so that my updated logo won't need to live in two different places. This is likely going to keep getting stranger as I move the README.md file away from the gater vignette (since the README has fewer requirements on it than the vignette, it might develop in a new direction as the project moves forward).