Delete README.html
Not sure what this html is about but I don't think I need it? Going to delete it and hopefully nothing bad will happen. I think it came from back when I was knitting the vignette both to vignettes and to the readme itself, so now that the README.md is separate from the vignette, this should be unimportant.